server/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.92.0
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.7
multidict==6.0.5
openai==0.27.0
pydantic==1.10.17
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.28.2
sniffio==1.3.1
starlette==0.25.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.26.19
uvicorn==0.30.6
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
yarl==1.9.4
